ai-web-app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
Flask-WTF
openai
python-dotenv
websockets

beautifulsoup4
requests
crawl4ai
